refactored UserRoleId fields. MVP Done.
