+++ shoppingcart/src/main/java/com/lambdaschool/shoppingcart/handlers/HelperFunctions.java
@@ -1,6 +1,5 @@
 package com.lambdaschool.shoppingcart.handlers;
 
-import com.lambdaschool.shoppingcart.models.User;
 import com.lambdaschool.shoppingcart.models.ValidationError;
 import org.springframework.stereotype.Component;
