--- shoppingcart/src/main/java/com/lambdaschool/shoppingcart/services/HelperFunctions.java
@@ -1,13 +1,0 @@
-package com.lambdaschool.shoppingcart.services;
-
-import com.lambdaschool.shoppingcart.models.User;
-import com.lambdaschool.shoppingcart.models.ValidationError;
-
-import java.util.List;
-
-public interface HelperFunctions
-{
-    List<ValidationError> getConstraintViolation(Throwable cause);
-
-    boolean isAuthorizedToMakeChange(String username);
-}
