+++ shoppingcart/src/main/java/com/lambdaschool/shoppingcart/services/HelperFunctionsService.java
@@ -1,0 +1,7 @@
+package com.lambdaschool.shoppingcart.services;
+
+
+public interface HelperFunctionsService
+{
+    boolean isAuthorizedToMakeChange(String username);
+}
